// frontend-service/src/interfaces/inputs.ts
export enum InputPickerType {
    ChooseFile = 'chooseFile',
    ChooseFromOptions = 'chooseFromOptions',
    ChooseAvatar = 'chooseAvatar',
    KartoffelProfile = 'kartoffelProfile',
}

export enum PermissionDialogMode {
    Create = 'create',
    Edit = 'edit',
    View = 'view',
}

export enum ItemTypes {
    FIELD = 'field',
    GROUP = 'group',
    STEP = 'step',
    PROPERTY = 'property',
}